refact: extrair listeners de durabilidade/quantidade/munição para equipment-field-listeners.mjs (actor-sheet.mjs)
