+++ module/sheets/actor-sheet.mjs
@@ -7,6 +7,7 @@ import { HeaderActions } from './actions/header-actions.mjs';
 import { HeaderStatusActions } from './actions/header-status-actions.mjs';
 import { HeaderListeners } from './listeners/header-listeners.mjs';
 import { AbilitiesListeners } from './listeners/abilities-listeners.mjs';
+import { EquipmentFieldListeners } from './listeners/equipment-field-listeners.mjs';
 import { ProficienciesActions } from './actions/proficiencies-actions.mjs';
 import { MoneyTradeActions } from './actions/money-trade-actions.mjs';
 import { InventoryActions } from './actions/inventory-actions.mjs';
@@ -356,15 +357,15 @@ export class CardiganSystemActorSheet extends api.HandlebarsApplicationMixin(
     this.#addExperienceListeners();
     
     // Adicionar event listeners para campos de durabilidade
-    this.#addDurabilityListeners();
-    this.#restoreDurabilityVisibility();
-    
+    EquipmentFieldListeners.addDurabilityListeners(this.element, this.actor, this);
+    EquipmentFieldListeners.restoreDurabilityVisibility(this.element, this);
+
     // Adicionar event listeners para campos de quantidade
-    this.#addQuantityListeners();
-    this.#syncQuantityInputWidths();
-    
+    EquipmentFieldListeners.addQuantityListeners(this.element, this.actor);
+    EquipmentFieldListeners.syncQuantityInputWidths(this.element);
+
     // Adicionar event listeners para campos de munição
-    this.#addAmmunitionListeners();
+    EquipmentFieldListeners.addAmmunitionListeners(this.element, this.actor);
     
     // Adicionar event listeners para campos dinâmicos de abilities
     AbilitiesListeners.initialize(this.element, this.actor);
@@ -1618,443 +1619,6 @@ export class CardiganSystemActorSheet extends api.HandlebarsApplicationMixin(
     // Update actor if there are any changes
     if (Object.keys(actorUpdates).length > 0) {
       await this.document.update(actorUpdates);
-    }
-  }
-
-  /********************
-   *
-   * Actor Override Handling
-   *
-   ********************/
-
-  /**
-   * Adiciona event listeners para campos de durabilidade de armas
-   * Implementa sincronização instantânea entre actor sheet e item sheet
-   */
-  #addDurabilityListeners() {
-    const html = this.element;
-    
-    // Encontrar todos os inputs de durabilidade de armas (atual e máxima)
-    const durabilityInputs = html.querySelectorAll('input[name*="durability"]');
-    
-    durabilityInputs.forEach(input => {
-      // Armazenar valor anterior para comparação
-      input.dataset.previousValue = input.value;
-      
-      // Usar apenas 'blur' para evitar processamento excessivo
-      input.addEventListener('blur', this.#handleDurabilityChange.bind(this));
-      
-      // Usar 'keydown' para capturar Enter
-      input.addEventListener('keydown', (e) => {
-        if (e.key === 'Enter') {
-          e.target.blur(); // Força o blur que irá processar a mudança
-        }
-      });
-    });
-
-    const durabilityToggleButtons = html.querySelectorAll('.item-durability .toggle-durability-display[data-action="toggleDurabilityDisplay"]');
-    durabilityToggleButtons.forEach(button => {
-      button.addEventListener('click', this.#handleDurabilityToggle.bind(this));
-    });
-  }
-
-  /**
-   * Alterna a visibilidade da durabilidade do item ao clicar no ícone
-   * @param {MouseEvent} event
-   */
-  #handleDurabilityToggle(event) {
-    event.preventDefault();
-    event.stopPropagation();
-
-    const toggleButton = event.currentTarget;
-    const itemRow = toggleButton.closest('li.item[data-item-id]');
-    const durabilityContainer = toggleButton.closest('.item-durability');
-
-    if (!itemRow || !durabilityContainer) return;
-
-    const itemId = itemRow.dataset.itemId;
-    if (!itemId) return;
-
-    const isExpanded = durabilityContainer.classList.toggle('is-durability-expanded');
-
-    if (isExpanded) {
-      this.durabilityExpandedItems.add(itemId);
-    } else {
-      this.durabilityExpandedItems.delete(itemId);
-    }
-  }
-
-  /**
-   * Restaura o estado de visibilidade da durabilidade após re-render
-   */
-  #restoreDurabilityVisibility() {
-    const html = this.element;
-    if (!html) return;
-
-    html.querySelectorAll('li.item[data-item-id] .item-durability').forEach(container => {
-      const itemRow = container.closest('li.item[data-item-id]');
-      const itemId = itemRow?.dataset.itemId;
-      if (itemId && this.durabilityExpandedItems.has(itemId)) {
-        container.classList.add('is-durability-expanded');
-      } else {
-        container.classList.remove('is-durability-expanded');
-      }
-    });
-  }
-
-  /**
-   * Manipula mudanças nos campos de durabilidade com debounce e validação
-   * @param {Event} event 
-   */
-  async #handleDurabilityChange(event) {
-    const input = event.target;
-    const name = input.name;
-    const currentValue = input.value.trim();
-    const previousValue = input.dataset.previousValue;
-    
-    // Se o valor não mudou, não faz nada
-    if (currentValue === previousValue) return;
-    
-    // Se o campo está vazio, não processa ainda (permite edição)
-    if (currentValue === '') {
-      console.log('[CARDIGAN] Campo vazio, aguardando valor...');
-      return;
-    }
-    
-    // Extrair o ID do item e o campo específico
-    const match = name.match(/^items\.([^.]+)\.system\.durability\.(.+)$/);
-    if (!match) return;
-    
-    const [, itemId, field] = match;
-    let value = parseInt(currentValue);
-    
-    // Validar se é um número válido
-    if (isNaN(value) || value < 0) {
-      console.log('[CARDIGAN] Valor inválido, restaurando valor anterior');
-      input.value = previousValue;
-      return;
-    }
-    
-    // Buscar o item para validação
-    const item = this.actor.items.get(itemId);
-    if (!item) return;
-    
-    // Validar limites baseados no tipo de campo
-    if (field === 'current') {
-      const maxDurability = item.system.durability.max;
-      if (value > maxDurability) {
-        console.log(`[CARDIGAN] Valor ${value} excede máximo ${maxDurability}, ajustando`);
-        value = maxDurability;
-        input.value = value;
-      }
-    } else if (field === 'max') {
-      // Durabilidade máxima deve ser pelo menos 1
-      if (value < 1) {
-        console.log(`[CARDIGAN] Durabilidade máxima deve ser pelo menos 1, ajustando`);
-        value = 1;
-        input.value = value;
-      }
-      
-      // Se a durabilidade atual for maior que a nova máxima, ajustar
-      const currentDurability = item.system.durability.current;
-      if (currentDurability > value) {
-        console.log(`[CARDIGAN] Ajustando durabilidade atual de ${currentDurability} para ${value}`);
-        await item.update({
-          'system.durability.current': value,
-          'system.durability.max': value
-        });
-        input.dataset.previousValue = value.toString();
-        return;
-      }
-    }
-    
-    console.log(`[CARDIGAN] Atualizando durabilidade - Item: ${itemId}, Campo: ${field}, Valor: ${value}`);
-    
-    // Preparar a atualização
-    const updateData = {};
-    updateData[`system.durability.${field}`] = value;
-    
-    try {
-      // Atualizar o item diretamente (sem usar o form submit para evitar conflito)
-      await item.update(updateData);
-      
-      // Atualizar o valor anterior para a próxima comparação
-      input.dataset.previousValue = value.toString();
-      
-      console.log(`[CARDIGAN] Durabilidade atualizada com sucesso para ${item.name}`);
-    } catch (error) {
-      console.error('[CARDIGAN] Erro ao atualizar durabilidade:', error);
-      // Restaurar valor anterior em caso de erro
-      input.value = previousValue;
-    }
-  }
-
-  /**
-   * Adiciona event listeners para campos de quantidade de itens da backpack
-   * Implementa sincronização instantânea entre actor sheet e item sheet
-   */
-  #addQuantityListeners() {
-    const html = this.element;
-    
-    // Encontrar todos os inputs de quantidade dos itens da backpack
-    const quantityInputs = html.querySelectorAll('.backpack-table .item-quantity input[type="number"]');
-    
-    quantityInputs.forEach(input => {
-      // Armazenar valor anterior para comparação
-      input.dataset.previousValue = input.value;
-
-      // Ajustar largura conforme a quantidade atual e enquanto digita
-      this.#updateQuantityInputWidth(input);
-      input.addEventListener('input', () => this.#updateQuantityInputWidth(input));
-      
-      // Usar apenas 'blur' para evitar processamento excessivo
-      input.addEventListener('blur', this.#handleQuantityChange.bind(this));
-    });
-  }
-
-  /**
-   * Ajusta o tamanho da fonte do input de XP com base na quantidade de dígitos.
-   */
-  #setupExperienceInputFontSize() {
-    const xpInput = this.element?.querySelector("input[name='system.experience.current']");
-    if (!(xpInput instanceof HTMLInputElement)) return;
-
-    const syncXPFontSize = () => {
-      const parsedValue = Number.parseInt(xpInput.value, 10);
-      const digits = Number.isNaN(parsedValue)
-        ? String(xpInput.value || '').replace(/\D/g, '').length
-        : String(Math.abs(parsedValue)).length;
-
-      xpInput.classList.toggle('is-three-digits', digits >= 3);
-    };
-
-    syncXPFontSize();
-    xpInput.addEventListener('input', syncXPFontSize);
-    xpInput.addEventListener('change', syncXPFontSize);
-    xpInput.addEventListener('blur', syncXPFontSize);
-  }
-
-  /**
-   * Adiciona event listeners para o campo de XP atual do header.
-   */
-  #addExperienceListeners() {
-    const xpInput = this.element?.querySelector("input[name='system.experience.current']");
-    if (!(xpInput instanceof HTMLInputElement)) return;
-
-    const clampXPValue = () => {
-      if (xpInput.value === '') return;
-
-      const parsedValue = Number.parseInt(xpInput.value, 10);
-      if (Number.isNaN(parsedValue)) {
-        xpInput.value = '0';
-        return;
-      }
-
-      const clampedValue = Math.max(0, Math.min(100, parsedValue));
-      xpInput.value = String(clampedValue);
-    };
-
-    clampXPValue();
-    xpInput.addEventListener('input', clampXPValue);
-    xpInput.addEventListener('blur', clampXPValue);
-  }
-
-  /**
-   * Ajusta a largura do input de quantidade com base na quantidade de dígitos.
-   * @param {HTMLInputElement} input - O input de quantidade.
-   */
-  #updateQuantityInputWidth(input) {
-    if (!input) return;
-
-    const digits = String(Math.abs(parseInt(input.value) || 0)).length;
-    const widthCh = digits === 1 ? 1.5 : digits === 2 ? 2.5 : Math.max(1, Math.min(3, digits));
-    input.style.width = `${widthCh}ch`;
-  }
-
-  /**
-   * Sincroniza a largura de todos os inputs de quantidade do inventário.
-   */
-  #syncQuantityInputWidths() {
-    const html = this.element;
-    const quantityInputs = html.querySelectorAll('.backpack-table .item-quantity input[type="number"]');
-    quantityInputs.forEach((input) => this.#updateQuantityInputWidth(input));
-  }
-
-  /**
-   * Manipula mudanças nos campos de quantidade dos itens da backpack
-   * @param {Event} event - O evento blur do input
-   */
-  async #handleQuantityChange(event) {
-    const input = event.target;
-    const name = input.name;
-    const value = parseInt(input.value) || 0;
-    const previousValue = input.dataset.previousValue;
-    
-    // Se o valor não mudou, não fazer nada
-    if (value.toString() === previousValue) return;
-    
-    console.log(`[CARDIGAN] Quantidade alterada: ${previousValue} → ${value}`);
-    
-    // Extrair o ID do item do nome do campo (formato: items.itemId.system.quantity)
-    const match = name.match(/^items\.([^.]+)\.system\.quantity$/);
-    if (!match) {
-      console.warn('[CARDIGAN] Nome do campo de quantidade inválido:', name);
-      return;
-    }
-    
-    const itemId = match[1];
-    const item = this.actor.items.get(itemId);
-    
-    if (!item) {
-      console.warn('[CARDIGAN] Item não encontrado:', itemId);
-      return;
-    }
-    
-    // Validar se é um dos tipos permitidos para edição de quantidade
-    const allowedQuantityTypes = ['item-comum', 'item-municao', 'item-consumivel', 'item-ingredient', 'arma', 'armadura'];
-    if (!allowedQuantityTypes.includes(item.type)) {
-      console.warn('[CARDIGAN] Tipo de item não permitido para edição de quantidade:', item.type);
-      return;
-    }
-    
-    // Validar valor mínimo por tipo
-    const minQuantity = ['arma', 'armadura'].includes(item.type) ? 1 : 0;
-    const finalValue = Math.max(minQuantity, value);
-    if (finalValue !== value) {
-      console.log(`[CARDIGAN] Valor ${value} ajustado para mínimo ${minQuantity}`);
-      input.value = finalValue;
-    }
-    
-    // Preparar a atualização
-    const updateData = {
-      'system.quantity': finalValue
-    };
-    
-    try {
-      // Atualizar o item diretamente
-      await item.update(updateData);
-      
-      // Atualizar o valor anterior para a próxima comparação
-      input.dataset.previousValue = finalValue.toString();
-      
-      console.log(`[CARDIGAN] Quantidade atualizada com sucesso para ${item.name}: ${finalValue}`);
-    } catch (error) {
-      console.error('[CARDIGAN] Erro ao atualizar quantidade:', error);
-      // Restaurar valor anterior em caso de erro
-      input.value = previousValue;
-    }
-  }
-
-  /**
-   * Adiciona event listeners para campos de munição de armas
-   * Implementa sincronização instantânea entre actor sheet e item sheet
-   */
-  #addAmmunitionListeners() {
-    const html = this.element;
-    
-    // Encontrar todos os inputs de munição de armas (atual e máxima)
-    const ammunitionInputs = html.querySelectorAll('input[name*="ammunition"]');
-    
-    ammunitionInputs.forEach(input => {
-      // Armazenar valor anterior para comparação
-      input.dataset.previousValue = input.value;
-      
-      // Usar apenas 'blur' para evitar processamento excessivo
-      input.addEventListener('blur', this.#handleAmmunitionChange.bind(this));
-      
-      // Usar 'keydown' para capturar Enter
-      input.addEventListener('keydown', (e) => {
-        if (e.key === 'Enter') {
-          e.target.blur(); // Força o blur que irá processar a mudança
-        }
-      });
-    });
-  }
-
-  /**
-   * Manipula mudanças nos campos de munição com debounce e validação
-   * @param {Event} event 
-   */
-  async #handleAmmunitionChange(event) {
-    const input = event.target;
-    const name = input.name;
-    const currentValue = input.value.trim();
-    const previousValue = input.dataset.previousValue;
-    
-    // Se o valor não mudou, não faz nada
-    if (currentValue === previousValue) return;
-    
-    // Se o campo está vazio, não processa ainda (permite edição)
-    if (currentValue === '') {
-      console.log('[CARDIGAN] Campo vazio, aguardando valor...');
-      return;
-    }
-    
-    // Extrair o ID do item e o campo específico
-    const match = name.match(/^items\.([^.]+)\.system\.ammunition\.(.+)$/);
-    if (!match) return;
-    
-    const [, itemId, field] = match;
-    let value = parseInt(currentValue);
-    
-    // Validar se é um número válido
-    if (isNaN(value) || value < 0) {
-      console.log('[CARDIGAN] Valor inválido, restaurando valor anterior');
-      input.value = previousValue;
-      return;
-    }
-    
-    // Buscar o item para validação
-    const item = this.actor.items.get(itemId);
-    if (!item) return;
-    
-    // Validar limites baseados no tipo de campo
-    if (field === 'current') {
-      const maxAmmunition = item.system.ammunition.max;
-      if (value > maxAmmunition) {
-        console.log(`[CARDIGAN] Valor ${value} excede máximo ${maxAmmunition}, ajustando`);
-        value = maxAmmunition;
-        input.value = value;
-      }
-    } else if (field === 'max') {
-      // Munição máxima deve ser pelo menos 0
-      if (value < 0) {
-        console.log(`[CARDIGAN] Munição máxima deve ser pelo menos 0, ajustando`);
-        value = 0;
-        input.value = value;
-      }
-      
-      // Se a munição atual for maior que a nova máxima, ajustar
-      const currentAmmunition = item.system.ammunition.current;
-      if (currentAmmunition > value) {
-        console.log(`[CARDIGAN] Ajustando munição atual de ${currentAmmunition} para ${value}`);
-        await item.update({
-          'system.ammunition.current': value,
-          'system.ammunition.max': value
-        });
-        input.dataset.previousValue = value.toString();
-        return;
-      }
-    }
-    
-    console.log(`[CARDIGAN] Atualizando munição - Item: ${itemId}, Campo: ${field}, Valor: ${value}`);
-    
-    // Preparar a atualização
-    const updateData = {};
-    updateData[`system.ammunition.${field}`] = value;
-    
-    try {
-      // Atualizar o item diretamente (sem usar o form submit para evitar conflito)
-      await item.update(updateData);
-      
-      // Atualizar o valor anterior para a próxima comparação
-      input.dataset.previousValue = value.toString();
-      
-      console.log(`[CARDIGAN] Munição atualizada com sucesso para ${item.name}`);
-    } catch (error) {
-      console.error('[CARDIGAN] Erro ao atualizar munição:', error);
-      // Restaurar valor anterior em caso de erro
-      input.value = previousValue;
     }
   }
